Change things to longs in serial MPD to make code work for edge cases

+++ src/mpd/SerialMinimumPairwiseDistance.java
@@ -4,11 +4,11 @@ public class SerialMinimumPairwiseDistance implements MinimumPairwiseDistance {
 
     @Override
     public long minimumPairwiseDistance(int[] values) {
-        long result = Integer.MAX_VALUE;
+        long result = Long.MAX_VALUE;
         for (int i = 0; i < values.length; ++i) {
             for (int j = 0; j < i; ++j) {
                 // Gives us all the pairs (i, j) where 0 <= j < i < values.length
-                long diff = Math.abs(values[i] - values[j]);
+                long diff = Math.abs((long)values[i] - values[j]);
                 if (diff < result) {
                     result = diff;
                 }
